Create a votepanel component to separate functionality in the personinformation component
