+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/index.js":"1","/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/components/app.js":"2","/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/components/personinformation.js":"3"},{"size":210,"mtime":1607741197281,"results":"4","hashOfConfig":"5"},{"size":178,"mtime":1607743439659,"results":"6","hashOfConfig":"5"},{"size":2571,"mtime":1607748923827,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1fz0lh",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/index.js",[],"/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/components/app.js",[],"/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/components/personinformation.js",["14"],{"ruleId":"15","severity":1,"message":"16","line":20,"column":3,"nodeType":"17","messageId":"18","endLine":20,"endColumn":14},"no-console","Unexpected console statement.","MemberExpression","unexpected"]
+[{"/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/index.js":"1","/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/components/app.js":"2","/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/components/personinformation.js":"3","/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/components/votepanel.js":"4"},{"size":210,"mtime":1607741197281,"results":"5","hashOfConfig":"6"},{"size":178,"mtime":1607743439659,"results":"7","hashOfConfig":"6"},{"size":2571,"mtime":1607748923827,"results":"8","hashOfConfig":"6"},{"size":2134,"mtime":1607749415211,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1fz0lh",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"16","messages":"17","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/index.js",[],"/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/components/app.js",[],"/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/components/personinformation.js",["18"],"/media/mateo/Backup 4TB/Disco/Mm/Microverse/Interview tests/zemoga/zemoga-test/src/components/votepanel.js",["19"],{"ruleId":"20","severity":1,"message":"21","line":20,"column":3,"nodeType":"22","messageId":"23","endLine":20,"endColumn":14},{"ruleId":"20","severity":1,"message":"21","line":21,"column":3,"nodeType":"22","messageId":"23","endLine":21,"endColumn":14},"no-console","Unexpected console statement.","MemberExpression","unexpected"]
