+++ src/components/personinformation.js
@@ -5,39 +5,10 @@
 /* eslint-disable react/self-closing-comp */
 /* eslint-disable arrow-body-style */
 
-import { faThumbsDown, faThumbsUp } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { useState } from 'react';
 import PersonInfoStyles from '../assets/stylesheets/personinfo.module.css';
+import VotePanel from './votepanel';
 
 const PersonInformation = () => {
-  const defaultVote = {
-    voted: false,
-    vote: null,
-  };
-
-  const [vote, setVote] = useState(defaultVote);
-  console.log(vote);
-
-  const select = event => {
-    const voteSelect = event.target;
-
-    const badButton = document.getElementById('Bad');
-    const goodButton = document.getElementById('Good');
-
-    badButton.classList.remove(PersonInfoStyles.Selected);
-    goodButton.classList.remove(PersonInfoStyles.Selected);
-
-    voteSelect.classList.add(PersonInfoStyles.Selected);
-    setVote({ ...vote, vote: voteSelect.id });
-  };
-
-  const submitVote = () => {
-    if (vote.vote !== null) {
-      setVote({ ...vote, voted: true });
-    }
-  };
-
   return (
     <div className={PersonInfoStyles.Container}>
       <h1>Person Name</h1>
@@ -47,38 +18,7 @@ const PersonInformation = () => {
         Quisque non elit massa. Donec quis turpis nisi.
         Mauris ultricies eget nisi id iaculis.
       </p>
-
-      <div className={PersonInfoStyles.Voting}>
-        { !vote.voted
-          ? (
-            <div
-              id="Good"
-              onClick={select}
-              className={`${PersonInfoStyles.VoteItem} ${PersonInfoStyles.Good}`}
-            >
-              <FontAwesomeIcon icon={faThumbsUp} />
-            </div>
-          )
-          : null}
-        { !vote.voted
-          ? (
-            <div
-              id="Bad"
-              onClick={select}
-              className={`${PersonInfoStyles.VoteItem} ${PersonInfoStyles.Bad}`}
-            >
-              <FontAwesomeIcon icon={faThumbsDown} />
-            </div>
-          )
-          : null}
-        <button
-          onClick={submitVote}
-          type="button"
-          className={`${PersonInfoStyles.VoteItem} ${PersonInfoStyles.Button}`}
-        >
-          {!vote.voted ? 'Vote Now' : 'Vote Again'}
-        </button>
-      </div>
+      <VotePanel />
     </div>
   );
 };
